requirements.txt
Django==1.8
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
